Make Invaders More Agressive
It seemed to me that invading creeps would only attack creeps that
attacked them (what I understand to be the hostile creeps).  When
perfoming an invastion I flelt it would be better to have then attack
anything that is not a white-list player's.  This will include creeps
that cannot attack (such as miners, claimers and upgraders) which helps
keep contested rooms free.