Add Basic Prospector
This first step combines some of the miner and privateer to create a
prospector.  This is a creep that performs dedicated mining in a remote
room.

The present state of the code does nothine more complciated than assign
a single prospector to a room which is set to be exploitted (Red+Greep
flag).  Each owned room generates one prospector at present.

When the prospector reaches a remote room it will either perform drop
mining or stash the energy in a container if one is provided.  The
problem with the container method is that privateers do not yet collect
energy from a friendly contianer for use.  I will either do that next as
a stop gap or begin work on a dedicated mule type.  This will basically
be a hauler set to work in a remote room and bring energy home.